:recycle: aws기능, 보컬분석 기능 모듈화
