+++ MyV/modal/templates/modal/analyze_result.html
@@ -52,7 +52,7 @@
                     <p id="report_paragraph">보고서 내용</p>
                 </div>
                 <h3 class="report_score">김동국님의 총 점수 : </h3>
-                <h3 class="report_score">80/100 점 </h3>
+                <h3 class="report_score">{{ score }}/100 점 </h3>
                 <p>안녕하세요 <b>김동국</b>님!</p>
                 <p>김동국님이 부른 부분은 음원의 <b>{{ start_t }}</b>초 부터 입니다.</p>
                 <p>최고음은 <b>{{ max }}</b> 이며
